🚚(cms) rename "landing pages" to "service pages"
landing page is not that explicit, the homepage is also a landing page.
now we understand better the usecase
