+++ content/.examples/service-page.yml
@@ -1,7 +1,6 @@
 title: Notepad
 hero:
-  title:
-    '**Notepad de l’État** <br>, Le meilleur moyen d’écrire <br> et partager
+  title: '**Notepad de l’État**, <br>Le meilleur moyen d’écrire <br>et partager
     votre savoir'
   subtitle: 'Écrivez, collaborez, diffusez : le Notepad est le point central de
     vos prises de notes.'
