+++ src/cms/collections/index.ts
@@ -1,11 +1,11 @@
 import { collection as homepageFile } from './homepage'
 import { collection as communsFile } from './communs'
-import { collection as landingPageCollection } from './landing-page'
+import { collection as servicePageCollection } from './service-page'
 import { collection as contentPageCollection } from './content-page'
 
 const collections = {
   fileCollections: [homepageFile, communsFile],
-  folderCollections: [landingPageCollection, contentPageCollection],
+  folderCollections: [servicePageCollection, contentPageCollection],
 }
 
 export default collections
